添加数据的保存

+++ migrations/20230131110054_temperature.sql
@@ -25,18 +25,18 @@ DROP TABLE IF EXISTS `electricity`;
 
 CREATE TABLE `electricity`
 (
-    remainder DOUBLE NOT NULL,
-    unit    INT NOT NULL,
-    created_at  Timestamp DEFAULT CURRENT_TIMESTAMP,
-    updated_at  Timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+    remainder  DOUBLE NOT NULL,
+    unit       DOUBLE NOT NULL,
+    created_at Timestamp DEFAULT CURRENT_TIMESTAMP,
+    updated_at Timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
 
 DROP TABLE IF EXISTS `water`;
 
 CREATE TABLE `water`
 (
-    remainder DOUBLE NOT NULL,
-    unit    INT NOT NULL,
-    created_at  Timestamp DEFAULT CURRENT_TIMESTAMP,
-    updated_at  Timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+    remainder  DOUBLE NOT NULL,
+    unit       DOUBLE NOT NULL,
+    created_at Timestamp DEFAULT CURRENT_TIMESTAMP,
+    updated_at Timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
